Remove redundant auth helper type declarations
