--- app/services/authHelper/types.ts
@@ -1,38 +1,0 @@
-export type SignInValues = {
-  email: string;
-  password: string;
-};
-
-export type SignUpValues = {
-  email: string;
-  password: string;
-};
-
-export type VerifyValues = {
-  email: string;
-  verificationCode: string;
-};
-
-export type SignUpResult = {
-  isSignUpComplete?: boolean;
-  nextStep?: {
-    signUpStep?: string;
-    codeDeliveryDetails?: {
-      attributeName?: string;
-      deliveryMedium?: string;
-      destination?: string;
-    };
-  };
-  userId?: string;
-  username?: string;
-};
-
-export type SignInResult = {
-  isSignedIn: boolean;
-  nextStep?: {
-    signInStep?: string;
-    additionalInfo?: Record<string, unknown>;
-  };
-  userId?: string;
-  username?: string;
-};
